% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbbdata-package.R
\docType{package}
\name{cbbdata-package}
\alias{cbbdata}
\alias{cbbdata-package}
\title{cbbdata: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
