% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_all_metrics.R
\name{cbd_all_metrics}
\alias{cbd_all_metrics}
\title{Get T-Rank, WAB, KP, BPI, SOR/SOS, and NET in one tibble}
\usage{
cbd_all_metrics(...)
}
\description{
Get T-Rank, WAB, KP, BPI, SOR/SOS, and NET in one tibble
}
\examples{
\donttest{try(cbd_all_metrics())}
}
