% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_ap_rankings.R
\name{cbd_ap_rankings}
\alias{cbd_ap_rankings}
\title{AP Poll History}
\usage{
cbd_ap_rankings(...)
}
\arguments{
\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function. Else, you can filter on  \code{team}, \code{conf}, or \code{year}.}
}
\description{
Pulls AP poll history from 1949-Present
}
\examples{
\donttest{try(cbd_ap_polls(team = "Duke"))}
}
