% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_bpi_ratings.R
\name{cbd_bpi_ratings}
\alias{cbd_bpi_ratings}
\title{BPI Ratings and ESPN Projections}
\usage{
cbd_bpi_ratings(...)
}
\arguments{
\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function. Else, you can filter on  \code{team} or \code{conf}.
|}
}
\description{
Pulls current BPI ratings and ESPN season and tournament projections.
}
\examples{
\donttest{try(cbd_bpi_ratings(conf = 'ACC'))}
}
