% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_create_account.R
\name{cbd_create_account}
\alias{cbd_create_account}
\title{Register for a CBBData Account}
\usage{
cbd_create_account(username, email, password, confirm_password)
}
\arguments{
\item{username}{Username}

\item{email}{Email}

\item{password}{Password}

\item{confirm_password}{Re-enter password}
}
\description{
Allows a user to create a CBBData API account to retrieve an API key
}
