% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_gt_logos.R
\name{cbd_gt_logos}
\alias{cbd_gt_logos}
\title{Render in Logo-Team HTML for Tables}
\usage{
cbd_gt_logos(
  data,
  team_column,
  logo_column = "team_logo",
  logo_type = "logo",
  logo_height = 25
)
}
\arguments{
\item{data}{Pass through your plotting data}

\item{team_column}{Indicate which column contains your \code{team} variable}

\item{logo_column}{Indicates the name of the HTML column}

\item{logo_type}{Indicate whether you want to plot logos ('logo', default) or
wordmarks ('wordmark')}

\item{logo_height}{The height of the logo or wordmark in the HTML string}
}
\value{
Returns data with an appended HTML column.
}
\description{
Appends a column with HTML that will render in team logos and names
for table plotting.
}
\details{
I really hate putting team logos and names in separate columns,
and this helper function will add team logos or wordmarks and create
the HTML code to render in logos and teams in the same column for tables.
If you are using \code{gt} to plot, you must use \code{fmt_markdown(team_logo)} to
render in the HTML. Set \code{logo_column} to your \code{team} variable if you wish
to overwrite your \code{team} column with the logo HTML.
}
\examples{
\dontrun{try(cbd_torvik_ratings(year=2023) \%>\% head() \%>\% select(team, conf) \%>\%
cbbdata::cbd_gt_logos(team, team) \%>\% gt::gt() \%>\% gt::fmt_markdown(team))}

}
