% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_kenpom_authorization.R
\name{cbd_kenpom_authorization}
\alias{cbd_kenpom_authorization}
\title{Authorize your account with KenPom}
\usage{
cbd_kenpom_authorization(password)
}
\arguments{
\item{password}{KenPom password, NOT CBBData password.}
}
\description{
Grants a user access to KenPom endpoints if they have an active KenPom
account.
}
\details{
Your CBBData account email must match your KenPom email. Pass through your
KenPom password. You will be asked to reauthorize on your KenPom account
expiration date.
}
