% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_kenpom_player_factors.R
\name{cbd_kenpom_player_factors}
\alias{cbd_kenpom_player_factors}
\title{KenPom Player Factors and Ranks}
\usage{
cbd_kenpom_player_factors(...)
}
\arguments{
\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function. Else, you can filter on `player`, `team`, `conf`, `year`,
or any other data column.}
}
\description{
Returns four factor statistics and national ranks from 2002-Present.
}
\details{
Requires KenPom authorization. You can authorize with
`cbd_kenpom_authorization`.
}
\examples{
\donttest{try(cbd_kenpom_player_factors(year = 2023))}
}
