% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_kenpom_ratings_archive.R
\name{cbd_kenpom_ratings_archive}
\alias{cbd_kenpom_ratings_archive}
\title{Day-to-day KenPom Ratings}
\usage{
cbd_kenpom_ratings_archive(...)
}
\description{
Historical KenPom ratings + rankings from 2012-Present. Requires KenPom
authorization. You can authorize with `cbd_kenpom_authorization`.
}
\details{
To load the complete data set, please set the `return_all` parameter to
`TRUE`. Else, your data will be filtered by your parameter selections.
}
