% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_kenpom_team_schedules.R
\name{cbd_kenpom_team_schedules}
\alias{cbd_kenpom_team_schedules}
\title{KenPom Team Schedules}
\usage{
cbd_kenpom_team_schedules(...)
}
\arguments{
\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function. Else, you can filter on  `team`, `year`, `date` or any
other data column.}
}
\description{
Pulls the KenPom team schedule page from 2002-Present.
}
\details{
Requires KenPom authorization. You can authorize with
`cbd_kenpom_authorization`.
}
