% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_login.R
\name{cbd_login}
\alias{cbd_login}
\title{Log-In to CBBData API}
\usage{
cbd_login(username = Sys.getenv("CBD_USER"), password = Sys.getenv("CBD_PW"))
}
\arguments{
\item{username}{Username}

\item{password}{Password}
}
\description{
Retrieves a user's CBBData API key.
}
\details{
If you are unsure on how to edit your \code{.Renviron} file, pass no arguments
through the function; it will walk you through the process and create a
persistent log-in. Else, pass your username and password through the function
to log-in for the current session only.
}
