% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_match_teams.R
\name{cbd_match_teams}
\alias{cbd_match_teams}
\title{Generate a team matching dictionary}
\usage{
cbd_match_teams()
}
\description{
A utility function that returns a named vector to be used as a matching
dictionary.
}
\examples{
\donttest{try(cbd_match_teams()['Duke Blue Devils'])}
}
