% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_net_archive.R
\name{cbd_net_archive}
\alias{cbd_net_archive}
\title{Daily NET Rankings}
\usage{
cbd_net_archive(...)
}
\arguments{
\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function. Else, you can filter on  \code{team}, \code{conf}, or \code{year}.}
}
\description{
Pulls daily NET rankings along with other metrics from the 2023-24 season.
}
\examples{
\donttest{try(cbd_net_archive(team = "Duke"))}
}
