% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_players.R
\name{cbd_players}
\alias{cbd_players}
\title{CBBData Player Database}
\usage{
cbd_players(...)
}
\arguments{
\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function. Else, you can filter on  `year`, `team`, or any other data
column.}
}
\description{
Pulls side-wide player IDs and meta data.
}
\examples{
\donttest{try(cbd_players())}
}
