% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_teams.R
\name{cbd_teams}
\alias{cbd_teams}
\title{CBBData Team Database}
\usage{
cbd_teams(...)
}
\arguments{
\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function.}
}
\description{
Pulls team matching dictionary + logos, colors, and wordmarks.
}
\examples{
\donttest{try(cbd_teams())}
}
