% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_conf_factors.R
\name{cbd_torvik_conf_factors}
\alias{cbd_torvik_conf_factors}
\title{Four Factor Conference Splits}
\usage{
cbd_torvik_conf_factors(
  year,
  venue = "all",
  game_type = "all",
  quad = "all",
  top = 0,
  start = NULL,
  end = NULL
)
}
\arguments{
\item{year}{Chosen year}

\item{venue}{Game venue ('all', 'home', 'away', 'neutral', 'road'). Defaults
to 'all'.}

\item{game_type}{Game type ('all', 'nc', 'conf', 'reg', 'post', 'ncaa').
Defaults to 'all'.}

\item{quad}{Quad. rank of game ('1', '2', '3', '4', 'all') and quads are
cumulative (games <= quad selection). Defaults to 'all', which is all games
(all games Q4 or "better"/"lower").}

\item{top}{Games against Top X opponents. Defaults to 0, which is "all"
games.}

\item{start}{Game start date (YYYYMMDD format).}

\item{end}{Game end date (YYYYMMDD format).}
}
\value{
Returns a tibble with four factor (and other!) data.
}
\description{
Split conference-wide four factor data on a number of variables.
}
\examples{
\donttest{try(cbd_torvik_conf_factors(2023, start = '20230101'))}

}
