% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_current_resume.R
\name{cbd_torvik_current_resume}
\alias{cbd_torvik_current_resume}
\title{Get Current Tournament Resumes}
\usage{
cbd_torvik_current_resume()
}
\value{
Returns a tibble with current resume data, including NET
}
\description{
Returns a tibble with current NCAA Tournament resume data
}
\examples{
\donttest{try(cbd_torvik_current_resume())}

}
