% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_game_box.R
\name{cbd_torvik_game_box}
\alias{cbd_torvik_game_box}
\title{Barttorvik Game Box Stats}
\usage{
cbd_torvik_game_box(...)
}
\arguments{
\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function. Else, you can filter on  \code{year}, \code{team}, \code{date}, \code{conf},
or any other data column.}
}
\description{
Pulls final game box stats from 2008-Present.
}
