% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_game_factors.R
\name{cbd_torvik_game_factors}
\alias{cbd_torvik_game_factors}
\title{Barttorvik Game Four Factors}
\usage{
cbd_torvik_game_factors(...)
}
\arguments{
\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function. Else, you can filter on  \code{year}, \code{team}, \code{date}, \code{conf},
or any other data column.}
}
\description{
Pulls final game four factors stats from 2008-Present.
}
