% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_game_prediction.R
\name{cbd_torvik_game_prediction}
\alias{cbd_torvik_game_prediction}
\title{T-Rank Game Predictions}
\usage{
cbd_torvik_game_prediction(team, opp, date = NULL, location = "H")
}
\arguments{
\item{team}{Team 1 in matchup (primary team)}

\item{opp}{Team 2 in matchup (opponent)}

\item{date}{Date of game and must be a consecutive string in YYYYMMDD format
(defaults to current day)}

\item{location}{Location of game relative to \code{team} ('H', 'A', or 'N'),
defaults to 'H' (home)}
}
\value{
Returns a tibble with game predictions.
}
\description{
Returns projected tempo, points per possesssion, total points scored,
and win percentage for any matchup at any location back to the 2014-15
season using Barttorvik's formulas and data.
}
\examples{
\donttest{try(cbd_torvik_game_prediction('Duke', 'Charlotte', '20221110'))}

}
