% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_game_stats.R
\name{cbd_torvik_game_stats}
\alias{cbd_torvik_game_stats}
\title{Barttorvik Full Game Stats}
\usage{
cbd_torvik_game_stats(...)
}
\arguments{
\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function. Else, you can filter on  \code{year}, \code{team}, \code{date}, \code{conf},
or any other data column.}
}
\description{
Barttorvik Full Game Stats
}
