% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_player_game.R
\name{cbd_torvik_player_game}
\alias{cbd_torvik_player_game}
\title{Barttorvik Player Game Logs}
\usage{
cbd_torvik_player_game(...)
}
\arguments{
\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function. Else, you can filter on  \code{year}, \code{team}, \code{player},or any
other data column.}
}
\description{
Pulls final player game logs from 2008-Present.
}
