% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_player_season.R
\name{cbd_torvik_player_season}
\alias{cbd_torvik_player_season}
\title{Barttorvik Player Season Averages}
\usage{
cbd_torvik_player_season(...)
}
\arguments{
\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function. Else, you can filter on  \code{year}, \code{team}, \code{player},or any
other data column.}
}
\description{
Pulls final player season averages from 2008-Present.
}
\examples{
\donttest{try(cbd_torvik_player_season(year = 2023, team = 'Duke'))}
}
