% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_player_split.R
\name{cbd_torvik_player_split}
\alias{cbd_torvik_player_split}
\title{Barttorvik Player Stat Splits}
\usage{
cbd_torvik_player_split(split, ...)
}
\arguments{
\item{split}{The data will default to splitting on game result. If you want
to split on a different level, you must specify. You can split on
\code{location}, \code{game_type}, \code{month}, or \code{game_result} (default).}

\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function. Else, you can filter on  \code{year}, \code{team}, \code{player}, or any
other data column.}
}
\description{
Pulls final player season averages by split from 2008-Present.
}
\examples{
\donttest{try(cbd_torvik_player_split(year = 2023, team = 'Duke', split = 'game_result'))}

}
