% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_resume_database.R
\name{cbd_torvik_resume_database}
\alias{cbd_torvik_resume_database}
\title{Get Selection Sunday Resumes}
\usage{
cbd_torvik_resume_database(
  n = 500,
  min_year = 2008,
  max_year = 2024,
  min_net = NULL,
  max_net = NULL,
  conf_group = NULL
)
}
\arguments{
\item{n}{Number of teams to return (defaults to 500).}

\item{min_year}{Minimum year to pull (YYYY).}

\item{max_year}{Maximum year to pull (YYYY).}

\item{min_net}{Minimum NET to filter.}

\item{max_net}{Maximum NET to filter.}

\item{conf_group}{Filter by conference or conference groups (accepts any
conference abbreviation, 'NCAA', 'Himajor', or 'mid').}
}
\value{
Returns a tibble with resume-based metrics.
}
\description{
Returns a tibble with various resume-based metrics through Selection Sunday.
Data runs from 2008-present.
}
\examples{
\donttest{try(cbd_torvik_resume_database(min_net = 1, max_net = 5))}

}
