% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_season_prediction.R
\name{cbd_torvik_season_prediction}
\alias{cbd_torvik_season_prediction}
\title{T-Rank Season Predictions}
\usage{
cbd_torvik_season_prediction(
  team,
  year,
  date = NULL,
  include_postseason = FALSE
)
}
\arguments{
\item{team}{Team to predict with.}

\item{year}{Year to run prediction on.}

\item{date}{Date of game and must be a consecutive string in YYYYMMDD format
(defaults to current day).}

\item{include_postseason}{A boolean representing whether conference
tournament and other postseason games should be included in the simulation
(defaults to FALSE).}
}
\value{
Returns a tibble with game predictions.
}
\description{
Returns projected tempo, points per possesssion, total points scored,
and win percentage for any team schedyle at any location back to the 2014-15
season using Barttorvik's formulas and data.
}
\examples{
\donttest{try(cbd_torvik_season_prediction('Duke', 2023, '20230101'))}

}
