% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_season_simulation.R
\name{cbd_torvik_season_simulation}
\alias{cbd_torvik_season_simulation}
\title{T-Rank Season Simulation}
\usage{
cbd_torvik_season_simulation(team, year, date, simulations = 10000)
}
\arguments{
\item{team}{Team to predict with.}

\item{year}{Year to run prediction on.}

\item{date}{Date of game and must be a consecutive string in YYYYMMDD format
(defaults to current day).}

\item{simulations}{An integer representing how many sims. to run (defaults to
10,000).}
}
\value{
Returns a tibble with simulation results.
}
\description{
Run a simulation to determine how many games a team is projected to win using
their schedule at any given date in the season.
}
\examples{
\donttest{try(cbd_torvik_season_simulation('Duke', 2023, '20230101'))}

}
