% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_similar_resumes.R
\name{cbd_torvik_similar_resumes}
\alias{cbd_torvik_similar_resumes}
\title{Similar Team Resumes}
\usage{
cbd_torvik_similar_resumes(
  team,
  year,
  conf_limit = "All",
  at_large_only = FALSE,
  current_resume = FALSE,
  net_weight = 1,
  resume_weight = 1,
  wab_weight = 1,
  elo_weight = 1,
  power_weight = 1
)
}
\arguments{
\item{team}{Chosen team}

\item{year}{Chosen year}

\item{at_large_only}{Boolean to filter on at-large teams only (defaults to
FALSE).}

\item{current_resume}{Boolean to filter on whether to use projected resume or
current resume (defaults to projected resume; FALSE).}

\item{net_weight}{Weight to assign NET ranking (defaults to 1).}

\item{resume_weight}{Weight to assign resume (defaults to 1).}

\item{wab_weight}{Weight to assign 'Wins Above Bubble' (WAB) (defaults to 1).}

\item{elo_weight}{Weight to assign ELO (defaults to 1).}

\item{power_weight}{Weight to assign power rating (defaults to 1).}

\item{conf_group}{Filter by conference or conference groups (accepts any
conference abbreviation, 'NCAA', 'Himajor', or 'mid').}
}
\value{
Returns a tibble with pre-tournament resumes
}
\description{
Find which teams share similar pre-tournament resumes (2008-present).
}
\examples{
\donttest{try(cbd_torvik_similar_resumes('Duke', 2023))}

}
