% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_team_history.R
\name{cbd_torvik_team_history}
\alias{cbd_torvik_team_history}
\title{Four Factor Team Histories}
\usage{
cbd_torvik_team_history(team)
}
\arguments{
\item{team}{team}
}
\value{
Returns a tibble with four factor history.
}
\description{
Returns a tibble with four factor team his
}
\examples{
\donttest{try(cbd_torvik_team_history(team = 'Duke'))}

}
