% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_team_schedule.R
\name{cbd_torvik_team_schedule}
\alias{cbd_torvik_team_schedule}
\title{Barttorvik Team Season Schedules}
\usage{
cbd_torvik_team_schedule(...)
}
\arguments{
\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function.}
}
\description{
Pulls season schedules from Barttorvik
}
\examples{
\donttest{try(cbd_torvik_team_schedule(team = 'Duke', year = 2023))}

}
