% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_team_split.R
\name{cbd_torvik_team_split}
\alias{cbd_torvik_team_split}
\title{Barttorvik Team Stat Splits}
\usage{
cbd_torvik_team_split(split, ...)
}
\arguments{
\item{split}{The data will default to splitting on game result. If you want
to split on a different level, you must specify. You can split on
\code{location}, \code{game_type}, \code{month}, or \code{game_result}.}

\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function. Else, you can filter on  \code{year}, \code{team}, \code{conf}, or any
other data column.}
}
\description{
Pulls final team season averages by split from 2008-Present.
}
\examples{
\donttest{try(cbd_torvik_team_split(year = 2023, split = 'month'))}

}
