% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbbdata-package.R
\docType{package}
\name{cbbdata-package}
\alias{cbbdata}
\alias{cbbdata-package}
\title{cbbdata: Interfaces the CBBData API}
\description{
Provides direct access to clean and tidy college basketball data using the CBBData API.
}
\author{
\strong{Maintainer}: Andrew Weatherman \email{andrew@aweatherman.com}

}
\keyword{internal}
