% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_kenpom_ratings.R
\name{cbd_kenpom_ratings}
\alias{cbd_kenpom_ratings}
\title{Year-End KenPom Ratings}
\usage{
cbd_kenpom_ratings(...)
}
\arguments{
\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function. Else, you can filter on  \code{team}, \code{conf}, \code{year}, or any
other data column.
|}
}
\description{
Pulls year-end KenPom ratings and adjusted efficiencies from 2001-Present.
}
\examples{
\donttest{try(cbd_kenpom_ratings(year = 2023))}
}
