% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_kenpom_ratings_archive.R
\name{cbd_kenpom_ratings_archive}
\alias{cbd_kenpom_ratings_archive}
\title{Day-to-day KenPom Ratings}
\usage{
cbd_kenpom_ratings_archive(...)
}
\arguments{
\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function. Else, you can filter on  \code{team}, \code{conf}, \code{year}, or any
other data column.}
}
\description{
Historical KenPom ratings + rankings from 2012-Present.
}
\details{
Requires KenPom authorization. You can authorize with
\code{cbd_kenpom_authorization}.
}
\examples{
\donttest{try(cbd_kenpom_ratings_archive(year = 2023))}
}
