% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_ratings_archive.R
\name{cbd_torvik_ratings_archive}
\alias{cbd_torvik_ratings_archive}
\title{Day-to-day T-Rank Ratings}
\usage{
cbd_torvik_ratings_archive(...)
}
\arguments{
\item{...}{OPTIONAL. To load the complete data set, pass no arguments through
to the function. Else, you can filter on  \code{team}, \code{conf}, \code{year}, or any
other data column.}
}
\description{
Historical T-Rank ratings + rankings from 2015-Present.
}
\examples{
\donttest{try(cbd_torvik_ratings_archive(year = 2023, conf = 'ACC'))}
}
