% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_torvik_season_simulation.R
\name{cbd_torvik_season_simulation}
\alias{cbd_torvik_season_simulation}
\title{T-Rank Season Simulation}
\usage{
cbd_torvik_season_simulation(
  team,
  year,
  simulations = 10000,
  date = NULL,
  include_postseason = FALSE
)
}
\arguments{
\item{team}{Team to predict with.}

\item{year}{Year to run prediction on.}

\item{simulations}{An integer representing how many sims. to run (defaults to
10,000).}

\item{date}{Date of game and must be a consecutive string in YYYYMMDD format
(defaults to current day).}

\item{include_postseason}{A boolean representing whether conference
tournament and other postseason games should be included in the simulation
(defaults to FALSE).}
}
\value{
Returns a tibble with simulation results.
}
\description{
Run a simulation to determine how many games a team is projected to win using
their schedule at any given date in the season.
}
\examples{
\donttest{try(cbd_torvik_season_simulation('Duke', 2023, date = '20230101'))}

}
