% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_theme_athletic.R
\name{gt_theme_athletic}
\alias{gt_theme_athletic}
\title{The Athletic `gt` Table Theme}
\usage{
gt_theme_athletic(gt_object, ...)
}
\arguments{
\item{gt_object}{An existing gt table object of class `gt_tbl`}

\item{...}{Optional additional arguments to `gt::table_options()`}
}
\value{
Returns data with an appended HTML column.
}
\description{
A theme for styling `gt` tables similar to The Athletic.
}
\examples{
\donttest{try(mtcars \%>\% head() \%>\% gt() \%>\% gt_theme_athletic())}

}
